add endpoints Api

+++ Fina.Core/Configuration.cs
@@ -3,7 +3,7 @@
     public static class Configuration
     {
         public const int DefaultStatusCode = 200;
-        public const long DefaultPageSize = 25;
-        public const long DefaultPageNumber = 1;
+        public const int DefaultPageSize = 25;
+        public const int DefaultPageNumber = 1;
     }
 }
